feat(lock): Improved restarting/failure semantics. (#2369)
Changes lock acquisition failure to stop the current branch, and fail
the pipeline when other branches complete. If one operation in parallel
is unable to acquire a lock, this would mean other operations would
complete normally (if they are able to lock) before the overall execution
fails.

This is preferable to one operation failing to acquire a lock leaving other
operations in a partially completed state. Additionally this makes the stage
restart story a bit more sensible - we can restart the failed Acquire Lock
stage and the pipeline will proceed, and we haven't failed a bunch of other
stages halfway through so the pipeline will proceed for any downstream join
points.

In the course of investigating, I discovered RestartableStage only existed
as a vestigal tail from the spring batch days, and deleted it.