feat(traffic-guards): consider capacity ratio (#2509)
* refactor(traffic-guards): express the condition in terms of capacity

The behavior should be essentially unchanged. The log message includes
more context, but Traffic Guards should still work as usual. This prepares
the way for a different condition based on capacity.

* refactor(tests): mostly cosmetic changes in TrafficGuardSpec

* use symbols when appropriate instead of copies of strings all over
  (such as "app-foo-v001")

* generate the test data in each test instead of mutating shared data

* feat(traffic-guard): bypass traffic guards if cluster currently has 0 capacity

It makes it impossible to delete an empty server group for instance,
which seems a bit pointless.

* feat(traffic-guards): consider capacity ratio

Instead of just enforcing capacity > 0, we want to compare the current
capacity and the future capacity. This is meant to prevent incidents where
a production cluster could accidentally go from 200 to 3 instances for
example.

* feat(traffic-guards): make capacity ratio configurable